refactored playlist to library and vice versa, also updated all references and tests, should be working now
