+++ contextproject-ms3/src/main/resources/views/library.fxml
@@ -9,31 +9,18 @@
 
 <TableView fx:controller="contextproject.controllers.LibraryController"
 	xmlns:fx="http://javafx.com/fxml" fx:id="tableView"
-	GridPane.hgrow = "ALWAYS"
-	GridPane.vgrow = "ALWAYS">
+	minWidth="220"
+	prefWidth="250">
 	<stylesheets>
 		<URL value="@../stylesheets/library.css" />
 	</stylesheets>
 	<columns>
-		<TableColumn text="Title" sortable="false">
+		<TableColumn text="Title" fx:id="nameColumn" sortable="false"
+		minWidth="220"
+		prefWidth="250">
 		  <cellValueFactory>
-        <PropertyValueFactory property="title" />
+        <PropertyValueFactory property="name" />
       </cellValueFactory>
 		</TableColumn>
-		<TableColumn text="Artist" sortable="false">
-		  <cellValueFactory>
-        <PropertyValueFactory property="artist" />
-      </cellValueFactory>
-		</TableColumn>
-		<TableColumn text="BPM" sortable="false">
-		  <cellValueFactory>
-        <PropertyValueFactory property="bpm" />
-      </cellValueFactory>
-		</TableColumn>
-		<TableColumn text="Key" sortable="false">
-		  <cellValueFactory>
-        <PropertyValueFactory property="key" />
-      </cellValueFactory>
-    </TableColumn>
 	</columns> 
 </TableView>
